backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
openai
google-generativeai